requirements.txt
PySide6==6.10.0
pyserial==3.5

